requirements.txt
Pillow==8.2.0
pyautogui==0.9.53
smtplib
pyinstaller